Hasta este punto está funcional el controlador de cita, hubieron una serie de modificaciones que pueden haber a futuro pero todo bien

+++ BackEstiloBGA/src/main/java/com/example/backestilobga/repositorio/CitaRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CitaRepositorio extends JpaRepository<Cita, Integer> {
+public interface CitaRepositorio extends JpaRepository<Cita, Long> {
 }
